da_sys_django/requirements.txt
django==3.2.5
mysql-connector-python==8.0.28
django-allauth==0.45.0
uwsgi==2.0.20
djangorestframework==3.12.4
django-extensions==3.1.3
django_redis==5.2.0
django-mysql==4.5.0
